+++ crates/lkjscript-compiler/src/hir/memory_plan/tests/fixtures.rs
@@ -58,7 +58,7 @@ pub(super) fn program(
 }
 
 #[allow(clippy::expect_used)]
-pub(super) fn product(id: u16, name: &str, fields: &[(&str, hir::Type)]) -> hir::ProductDefinition {
+pub(super) fn product(id: u64, name: &str, fields: &[(&str, hir::Type)]) -> hir::ProductDefinition {
     let identity = lkjscript_contracts::sha256(&id.to_be_bytes());
     hir::ProductDefinition {
         id: hir::ProductId::new(id),
@@ -69,7 +69,7 @@ pub(super) fn product(id: u16, name: &str, fields: &[(&str, hir::Type)]) -> hir:
             .iter()
             .enumerate()
             .map(|(index, (name, ty))| {
-                let source_order = u16::try_from(index).expect("fixture field order");
+                let source_order = u64::try_from(index).expect("fixture field order");
                 hir::ProductField {
                     identity: crate::source::product_field_identity(identity, name, source_order)
                         .expect("fixture product field identity"),
@@ -118,14 +118,14 @@ pub(super) fn enum_definition(
         .map(|(vi, (name, fields))| hir::EnumVariant {
             id: variant_id(id.wrapping_add(1 + u8::try_from(vi).unwrap_or(0))),
             name: name.into(),
-            source_order: u16::try_from(vi).unwrap_or(u16::MAX),
+            source_order: u64::try_from(vi).unwrap_or(u64::MAX),
             fields: fields
                 .into_iter()
                 .enumerate()
                 .map(|(fi, ty)| hir::EnumVariantField {
                     id: field_id(id.wrapping_add(32 + u8::try_from(fi).unwrap_or(0))),
                     name: format!("field-{fi}"),
-                    source_order: u16::try_from(fi).unwrap_or(u16::MAX),
+                    source_order: u64::try_from(fi).unwrap_or(u64::MAX),
                     ty,
                     indirect: false,
                 })
